Preserve fractional transfer timeouts

+++ lib/judges/commands/upload.rb
@@ -38,7 +38,7 @@ class Judges::Upload
     baza = BazaRb.new(
       opts['host'], opts['port'].to_i, opts['token'],
       ssl: opts['ssl'],
-      timeout: (opts['timeout'] || 30).to_i,
+      timeout: (opts['timeout'] || 30).to_f,
       loog: @loog,
       retries: (opts['retries'] || 3).to_i
     )
